feat(audio): port Apogee Multivoc audiolib; fix CACHE SPACE ALL LOCKED UP crash
- Compile full Multivoc stack (fx_man, multivoc, mv_mix, ll_man, pitch,
  dsl, user) via pre_build.py GLOB_RECURSE; exclude DOS/ISA drivers and
  OPL2/MIDI music sources that conflict or pull in duke3d.h globals
- dsl.c: FreeRTOS audio pump task on Core 1 priority 4; sleeps 23 ms per
  buffer period, mixes via MV_ServiceVoc under portMUX, pushes to I2S
  non-blocking (ticks_to_wait=0); DisableInterrupts/RestoreInterrupts
  mapped to same portMUX for voice-list atomicity with game task
- audiolib_stubs.c: reduced to MUSIC_* no-ops only; FX_* now from real lib
- i2s_audio: 11025 Hz stereo 16-bit, 8×256 DMA buffers (~370 ms);
  platform_audio_write duplicates mono samples L=R
- conio.h / dos.h: empty stubs for ESP32 build (multivoc.c includes them
  inside PLAT_DOS guards; headers still needed to satisfy preprocessor)
- Fix CACHE SPACE ALL LOCKED UP crash: move sound buffers out of tile
  cache into dedicated PSRAM heap (heap_caps_malloc MALLOC_CAP_SPIRAM);
  clearsoundlocks() now heap_caps_free()s non-playing sounds; tile cache
  capped at 256 KB to leave ~200 KB PSRAM headroom for sound buffers
- docs: add Audio system section covering stack, dsl.c design, mixing
  params, compiled files, music stubs, PSRAM layout, and the cache bug

+++ components/hub75_matrix/hub75_matrix.cpp
@@ -486,7 +486,7 @@ bool grp_title_splash_build_cache_if_needed(const char *grp_path, const char *ca
             if (grp_splash_load_lookup_title_drealms(fp, static_cast<long>(data_off), sizes[fi]))
                 have_lookup_pals = true;
             else
-                ESP_LOGW(TAG_GRP_SPLASH, "LOOKUP.DAT present but parse failed (size=%u)", sizes[fi]);
+                ESP_LOGW(TAG_GRP_SPLASH, "LOOKUP.DAT present but parse failed (size=%lu)", (unsigned long)sizes[fi]);
         }
         data_off += sizes[fi];
     }
